pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = ""
authors = ["shawn.kim <dun03191@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.10.5"
requests = "^2.32.3"
pandas = "^2.2.3"
yfinance = "^0.2.49"
ccxt = "^4.4.32"
ta-lib = "^0.5.1"
ipykernel = "^6.29.5"
pyyaml = "^6.0.2"
scipy = "^1.14.1"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
